pyproject.toml
[tool.poetry]
name = "AIEDU"
version = "0.1.0"
description = ""
authors = ["DrWhiteCloudAI <yenphan1@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "1.0.0"
tabulate = "0.9.0"
pyyaml = "6.0.1"
langchain = ">=0.0.352"
langchain-community = "^0.3.18"
pandas = "^2.2.3"
plotly = "^6.0.0"
langchain-experimental = "^0.3.4"
beautifulsoup4 = "^4.13.3"
chromadb = "^0.6.3"
langchain-chroma = "^0.2.2"
langchain-openai = "^0.3.7"
crewai = "^0.102.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
